requirements.txt
asgiref==3.6.0
Django==4.1.7
django-autoslug==1.9.9
django-environ==0.10.0
django-tinymce==3.6.1
dnspython==2.3.0
Pillow==9.4.0
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2-binary==2.9.6
pytz==2023.2
sqlparse==0.2.4
typing_extensions==4.5.0